get latest versions of git apps
* do not fail when installing chrome a second time

+++ roles/common/files/home.sh
@@ -1,8 +1,13 @@
 #!/bin/bash -e
 
 cd ~
-git init
-git remote add origin https://github.com/craigfurman/home.git
-git fetch
-git reset --hard origin/master
-git branch --set-upstream-to origin/master
+
+if [ -d ~/.git ]; then
+  git pull
+else
+  git init
+  git remote add origin https://github.com/craigfurman/home.git
+  git fetch
+  git reset --hard origin/master
+  git branch --set-upstream-to origin/master
+fi
